Make parseable in IE8
`catch` is a reserved word and IE8 hates when you use these as keys in objects unquoted. I'm not sure if this makes promises themselves work in IE8, but for my purposes it's enough (I'm rendering a dialog that IE8 is unsupported - adding quotes avoids a parsing error).

Related: https://github.com/RubenVerborgh/promiscuous/issues/22